Promote MPL_large_writev/readv return types to ssize_t.
We were originally returning "int" from these functions because the
idea was that they'd never read/write more than INT_MAX.  However, in
the case where the upper layer passes an IOV, where the first element
is smaller than INT_MAX, but the next element is larger than INT_MAX,
the total number of bytes sent might be larger than INT_MAX.

This was occassionally showing up as an error on Mac OS for large
message communication.

+++ src/mpl/include/mplsock.h
@@ -34,8 +34,8 @@ extern "C" {
 #endif
 /* *INDENT-OFF* */
 
-int MPL_large_writev(int fd, const struct iovec *iov, int iovcnt);
-int MPL_large_readv(int fd, const struct iovec *iov, int iovcnt);
+ssize_t MPL_large_writev(int fd, const struct iovec *iov, int iovcnt);
+ssize_t MPL_large_readv(int fd, const struct iovec *iov, int iovcnt);
 
 /* *INDENT-ON* */
 #if defined(__cplusplus)
